change interface to be around obj and arr
