+++ arr.go
@@ -1,0 +1,92 @@
+package jman
+
+import (
+	"fmt"
+	"slices"
+)
+
+type Arr []any
+
+func (a Arr) Equal(other any, optFuncs ...OptsFunc) error {
+	opts := EqualOptions{}
+
+	for _, o := range optFuncs {
+		o(&opts)
+	}
+
+	if err := opts.valid(); err != nil {
+		return fmt.Errorf("invalid options: %w", err)
+	}
+
+	act, err := New[Arr](other)
+	if err != nil {
+		return fmt.Errorf("invalid actual: %w", err)
+	}
+
+	a, err = normalize(a)
+	if err != nil {
+		return fmt.Errorf("expected is invalid json: %w", err)
+	}
+
+	differences := compareArrays(base, a, act, opts)
+	if len(differences) > 0 {
+		return fmt.Errorf("expected not equal to actual:\n%s", differences.Report())
+	}
+
+	return nil
+}
+
+func compareArrays(path string, expected, actual Arr, opts EqualOptions) Differences {
+	var diffs Differences
+	if len(expected) != len(actual) {
+		diffs = append(diffs, Difference{
+			prefix: Expected,
+			diff:   fmt.Sprintf("expected %d items - got %d items", len(expected), len(actual)),
+		})
+	}
+
+	if slices.Contains(opts.ignoreArrayOrder, path) {
+		comparedDiffs := compareArraysIgnoreOrder(path, expected, actual, opts)
+		diffs = append(diffs, comparedDiffs...)
+	} else {
+		comparedDiffs := compareArraysStrictOrder(path, expected, actual, opts)
+		diffs = append(diffs, comparedDiffs...)
+	}
+	return diffs
+}
+
+func compareArraysIgnoreOrder(path string, expected, actual Arr, opts EqualOptions) Differences {
+	var diffs Differences
+	for i, item := range expected {
+		found := slices.ContainsFunc(actual, func(v any) bool {
+			equal, _ := compareValues(fmt.Sprintf("%s.%d", path, i), item, v, opts)
+			return equal
+		})
+		if found {
+			continue
+		}
+
+		d := Difference{
+			prefix: Expected,
+			path:   fmt.Sprintf("%s.%d", path, i),
+			diff:   "not found in actual",
+		}
+		diffs = append(diffs, d)
+	}
+	return diffs
+}
+
+func compareArraysStrictOrder(path string, expected, actual Arr, opts EqualOptions) Differences {
+	var diffs Differences
+	for i, item := range expected {
+		if i >= len(actual) {
+			continue
+		}
+		equal, diff := compareValues(fmt.Sprintf("%s.%d", path, i), item, actual[i], opts)
+		if equal {
+			continue
+		}
+		diffs = append(diffs, diff)
+	}
+	return diffs
+}
